Use ints in fibonacci
I could use size_t. I can write packers to translate that to ints in the
json. And I could write a UInt type for R and other languages. But
ultimately, all I am doing is doubling the size of the number that can
be expressed. Hardly a solution. Since the fibonacci function is mostly
just for use in simple examples, I'll stick to simple ints for now.

+++ cpp_math.hpp
@@ -4,8 +4,7 @@
 #include <math.h>
 #include <vector>
 
-
-std::vector<size_t> fibonacci(size_t);
+std::vector<int> fibonacci(int);
 
 template <class A>
 A sum(std::vector<A> xs){
@@ -31,8 +30,8 @@ A var(std::vector<A> xs){
     return(diff / (xs.size() - 1));
 }
 
-std::vector<size_t> fibonacci(size_t n){
-    std::vector<size_t> xs;
+std::vector<int> fibonacci(int n){
+    std::vector<int> xs;
     if(n <= 0){
        return xs; 
     }
